move 'wear' action to 'garments' module
